+++ test/chat_web/live/main_live/index_test.exs
@@ -270,14 +270,6 @@ defmodule ChatWeb.MainLive.IndexTest do
       assert html =~ "Cargo sync activated"
       assert html =~ "Insert empty USB drive"
 
-      assert(
-        String.contains?(html, "2:00") or String.contains?(html, "1:59"),
-        "not found in #{html}"
-      )
-
-      Process.sleep(1000)
-      assert render(view) =~ ~r/1:5\d/
-
       %{socket: socket} = reload_view(%{view: view})
       CargoRoom.sync(socket.assigns.room.pub_key)
       Process.sleep(100)
@@ -325,7 +317,6 @@ defmodule ChatWeb.MainLive.IndexTest do
       html = render(view)
       assert html =~ "Cargo sync activated"
       assert html =~ "Insert empty USB drive"
-      assert html =~ "2:00" or html =~ "1:59"
     end
 
     test "is disabled for rooms with non-unique names", %{view: view} do