+++ CHANGELOG.md
@@ -13,6 +13,7 @@ straightforward as possible.
 ### Added
 
 ### Changed
+- Hide Cargo sync timer and increase the timeout to 5 minutes
 
 ### Fixed
 - feeds close error on mobile bugfix